This update fully supports https plus other improvements
The same version is running on VPS  https://abcsales.duckdns.org/ 24x7
When you run it locally , point your browser to https://localhost  and ignore the browser warnings, that the site is NOT SECURE, this is because the certificate does not correspond to you local site. Normal thing in a development environment.